Reduce the number of tries to match rules
The worst case is when a module build with state build comes in, there is no
rule defined for this state, which probably only has rules for state done and
ready. In this case, MTS will try to test the module build with every rule.

With this patch, the build state is checked before starting to match rules. And
rules with same build state are checked only, others with different build state
are skipped.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>