Migrate to fedora-messaging
This patch drops fedmsg and uses fedora-messaging instead. From now on,
MTS has to interact with two message buses, fedora-messaging running on
top of rabbitmq and internal UMB running on top of ActiveMQ. Both of
them supports AMQP protocol but in different versions. As a result, it
is impossible to interact with all of them with a single messaging
library. So, for fedora-messaing, MTS calls the API provided by
fedora-messaging itself, and rhmsg is called for internal messaging.
Setting config messaging_backend to switch back and forth.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>