Run mts in the foreground
The message_tagging_service runs as a python module. In the container
image, gunicorn is also run for prometheus metrics. This is a secondary
capability of the container image. The most important being the python
process that run MTS. By making it run in the foreground,
OpenShift/Kubernetes can properly determine if the pod needs to be
restarted in case MTS quits unexpectedly.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>